+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 - The Coal Generator now allows extracting items from any side (in case of fuels like Lava Buckets, which leave an empty Bucket behind)
 - Machines now keep their progress if they run out of power
 - MobGeneRegistry and GeneRequirementRegistry loggers are now debug instead of info
+- The non-empty DNA Helices and Plasmids in the creative mode tab are now after everything else